Add AssemblyContext::recompute_elem_solution_rate
Turns out we need to manually recomptue the elem_solution_rate
if we're giong to use interior_rate since we're doing the finite
differencing within the Physics and not outside the TimeSolver.
When we perturb the solution dofs, the solution rate doesn't see
it, so we add a hook to manually recompute it.